requirements.txt
pytest==7.1.2
requests==2.27.1
pydantic==1.9.0
selenium==4.1.3
allure-pytest==2.9.45
